requirements.txt
# Web Framework
Flask==2.3.3
gunicorn==21.2.0
Flask-Limiter>=3.5.0
redis>=5.0.1

# Machine Learning & Deep Learning
torch>=2.2.0
transformers==4.46.3
datasets==2.14.5
scikit-learn==1.3.2
accelerate>=0.26.0
evaluate==0.4.3

# Document Processing
pypdf==5.1.0
pytesseract==0.3.13
Pillow==10.0.0
docx2txt==0.8
openpyxl==3.1.2
striprtf==0.0.26
python-magic==0.4.27

# Data Processing & Utilities
joblib==1.3.2
faker==33.0.0
protobuf>=4.25.0
tiktoken>=0.5.0

# Testing
pytest==7.4.3
pytest-mock==3.12.0